Fix autostart not working in Raspbian with Chromium
In Raspbian, the IDs of windows opened in Chromium on start up can change. This causes activateTab to grab non-existent settings as the function passes the window ID of a tab to grabTabSettings instead of passing an explicit parameter. This fix simply adds an argument to activateTab for passing a window ID.